backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidian-emergent-tags"
version = "0.1.0"
description = "Système de tags émergents pour Obsidian"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Robin Tual"}
]
dependencies = [
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "hdbscan>=0.8.33",
    "networkx>=3.0",
    "python-frontmatter>=1.0.0",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0.0",
    "click>=8.1.0",
    "tqdm>=4.65.0",
    "spacy>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
emergent-tags = "backend.src.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]
